requirements.txt
pytest==7.4.4
pytest-mock==3.12.0
black==23.12.1
ruff==0.1.14
pyodbc==5.2.0
python-dotenv==1.0.1